requirements.txt
selenium==4.33.0
pytest==8.4.0
requests==2.32.4
sqlalchemy==2.0.41
allure-pytest==2.14.3
faker==37.4.0
allure-python-commons~=2.14.3
webdriver_manager
git+https://AntonRuskhih:glpat-KY81g4JLGxgjs1wr1CR6@gitlab.com/barbariki245/api-requests.git@AntonioBANdeRASS#egg=boba_framework
git+https://AntonRuskhih:glpat-KY81g4JLGxgjs1wr1CR6@gitlab.com/barbariki245/automation.framework.git@main#egg=automation_framework

